// src/app/core/interface/medicamentoEmpaque.interface.ts
export interface MedicamentoEmpaque {
  idMedicamentoEmpaque:       number;
  idMedicamento:              number;
  descripcionMedicamento:     string;
  idPresentacion:             number;
  descripcionPresentacion:    string;
  cantidad:                   number;
  precioVenta:                string;
  precioCompra:               string;
  esActivo:                   number;
}